Added dependency on rnacentraldb, removed code that is there now; reorganized things
