+++ src/test/scala/Dbrna16s.scala
@@ -1,6 +1,6 @@
 package era7bio.db.test
 
-import era7bio.db.rna16sDBRelease._
+import era7bio.db._
 import org.scalatest.FunSuite
 import better.files._
 
@@ -9,7 +9,7 @@ class Dbrna16sTest extends FunSuite {
 
   test("Process some sample files") {
 
-    generateRna16sDB.processSources(
+    rna16sDB.generateBundle.processSources(
       file"source.table.sample.tsv",
       file"output.table.sample.tsv".clear()
     )(file"source.sample.fasta",
