+++ src/test/scala/runBundles.scala
@@ -1,0 +1,34 @@
+package era7bio.db.test
+
+import ohnosequences.statika._, aws._
+import ohnosequences.awstools._, regions.Region._, ec2._, InstanceType._, autoscaling._, s3._
+
+import era7bio.db._
+import era7.defaults._
+
+
+case object rna16sDBRelease {
+
+  case object compat extends Compatible(
+    amznAMIEnv(
+      AmazonLinuxAMI(Ireland, HVM, InstanceStore),
+      javaHeap = 20 // in G
+    ),
+    rna16sDB.generateBundle,
+    generated.metadata.DbRna16s
+  )
+
+  def launch(user: AWSUser): List[String] = {
+    EC2.create(user.profile)
+      .runInstances(
+        amount = 1,
+        compat.instanceSpecs(
+          c3.x2large,
+          user.keypair.name,
+          Some(ec2Roles.projects.name)
+        )
+      )
+      .map { _.getInstanceId }
+  }
+
+}
